+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - new diagram type flowcharts.
 
+### Changed
+
+- renamed `Add-Relation` to `Add-ErRelation`.
+
 ## [0.2.0] - 2022-09-14
 
 ### Added